Testing interface moved to a separate file
