+++ packages/chain/statemanager/sm_snapshots/interface_t.go
@@ -1,0 +1,8 @@
+package sm_snapshots
+
+type SnapshotManagerTest interface {
+	SnapshotManager
+	SnapshotReady(SnapshotInfo)
+	IsSnapshotReady(SnapshotInfo) bool
+	SetAfterSnapshotCreated(func(SnapshotInfo))
+}
